Update and rename work16/index1.js to  work16/index1.js

+++ work16/index1.js
@@ -27,7 +27,7 @@ function animate(obj, option){
     }
     var obj = document.getElementById('box');
     obj.onclick = function(){
-        var left = parseFloat(Math.random()* 400)
-        var top = parseFloat(Math.random()* 600)
+        var left = parseInt(Math.random()* 400)
+        var top = parseInt(Math.random()* 600)
         animate(obj, {'left':left, 'top':top});
     };
